Update .vimrc, and add more documentation for it to the README.
.vimrc is now friendlier to reload, by unsetting the global autocmds
near the start.

Add settings for vim-autoclose. '<' is autoclosed for html-like files.

Add cursorline setting. The 99th column color is tweaked to match the cursorline color.

Add wildmenu and lazyredraw.

Matchpairs now matches '`', to mirror autoclose's groups.

Add <leader>ez to edit the .zshrc file in the user's home folder.

Various updates to comments in .vimrc.

Various updates for the above in the README.